Debug Console / Socket line Parsing
I added a debug console window that prints debug info incliuding socket
comm.  I added status labels to the GUI for cleaner messages to the
user.  I am currently working on the socket line buffer.  I am
implementing a queue to correctly split up the incoming response from
the server.  It is being modeled after the buffer on the server.
NOTE ** The buffer is not yet done.